feat(EMS-1175): moved update inactive applications to helpers and code and test fixes
